feat(rds): Support Multi AZ Database Clusters
This adds support for Multi AZ Database Clusters. These are the non
aurora based database clusters.

Creating a new `MultiAZDatabaseCluster` class was needed as Multi AZ
Clusters do not need to create their own instances in the CloudFormation
template. Many of the properties that were getting set on each
individual instance with `DatabaseCluster` need to get set on the
cluster itself rather than the instance.

fixes #25547