feat(secretsmanager): add support for rotateImmediatelyOnUpdate for secret rotation schedule (#25652)
When adding a rotation schedule to a secret, you can specify whether or not the secret should be rotated immediately. This is optional, and by default, the secret will be rotated immediately if rotateImmediatelyOnUpdate is undefined. This PR exposes the rotateImmediatelyOnUpdate property enabling the user to configure this as false if they do not want the secret to be rotated immediately. Configuring this property as being explicitly true will result in the secret being rotated immediately, which is the default behavior.

Closes #25365

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*