feat(lambda): lambda code assets are marked as deploy time assets (#25705)
Mark lambda assets as deploy time assets. This has no functionality for default synthesis, but if used in conjunction with the `AppStagingSynthesizer`, lambda assets will be marked with the `deploy-time/` prefix in the staging bucket and subject to lifecycle policies. This is because for lambda code, the s3 staging bucket is an intermediate step and the code is copied to lambda immediately. The code in s3 is no longer referenced after deploy, except when in a rollback scenario.

Since this is a no-op for normal synthesis, the only test that can be done is in `app-staging-synthesizer-alpha`.

Also I've deleted a duplicate test I found in `app-staging-synthesizer-alpha`.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*