feat(route53): HostedZone's default period at the end should be optional  (#25379)
new flag `addTrailingDot`

The addTrailingDot option provides users with the choice to add a trailing dot (.) to the given domain name. In DNS, appending a dot at the end of a domain name indicates that the domain name is a fully qualified domain name (FQDN).
This function returns the value as is if the domain name is already an FQDN or has a trailing dot, and otherwise treats it as having a trailing dot. The existing specification was rejecting domain names with trailing dots, but this modification makes it unnecessary, so the validation check has been removed.

Closes #22406

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*