fix(aws-cdk-lib): attribute `FindingsFilterListItems` on AWS::Macie::FindingsFilter does not work (#25778)
The attribute was removed in a recent CFN Spec update.
We had then patched the attribute and type back in to avoid a breaking change. 
Now I have confirmed with the service team that this attribute has actually never worked and the removal is indeed correct.
Update CDK to reflect the fix made in CloudFormation.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*