chore: prlinter should use HEAD instead of base commit (#25779)
The `sha` for the `pull_request_target` event is for the base branch (i.e. `main`) so we were checking the status from the last commit to `main`. We should instead get the `sha` from the event payload which contains the `head` sha.

This also fixes an issue processing the status event where we were trying to parse the latest commit from the status URL, but that value is always equal to `{sha}`. Instead get it from the `head` info.

Added some debug logs to help us see what is getting processed. 

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*