feat(events-targets): support enableExecuteCommand in EcsTask construct (#25639)
Users can run commands in or get a shell to a container running on an Amazon EC2 instance or on AWS Fargate with Amazon ECS Exec.

`FargateService` and `Ec2Service` in `aws-ecs` supports the flag to enable Amazon ECS Exec, `enableExecuteCommand`.
`EcsTask` in `aws-events-targets` also supports `enableExecuteCommand` in CloudFormation, but AWS CDK doesn't.
This feature supports `enableExecuteCommand` flag also for `EcsTask`.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*