fix(lambda-nodejs): cannot use .mts, .cts, and .cjs entry files (#25642)
This adds support for using TypeScript Module (`.mts`), TypeScript CommonJS (`.cts`) and JavaScript CommonJS (`.cjs`) entry files, as well as adding these to the default entry file search. 

This is a must for projects mixing ESM and CommonJS (quite common as not everything can/should be fully ESM yet) or where explicit file extension is desired.

I tried to follow the existing convention as best I could to keep it as simple as possible. I am not sure why the tests for the existing entry files are skipped as they seem to be working correctly, at least locally, so I enabled the new tests I added.

Closes #21635.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*