fix(cli): ENOENT during asset publishing
There was still a TOCTOU error in the file asset publishing, which could
lead to `ENOENT: no such file or directory` when assets were being published
in parallel.

The whole `if (fileExists) { delete; }` logic was actually not
necessary, as `fs.rename` will atomically overwrite existing files
already, so we can just call it directly.

Closes #25293.