fix(ecs): remove accidental duplication of cloudmap namespaces with service connect (#25891)
This PR should fix #25616, where service connect accidentally creates a duplicate HTTP namespace when a customer sets a service connect default namespace on the cluster. 

Closes #25616 

However, I think that a broader fix for this issue should include deprecation of the `namespace` parameter in `ServiceConnectProps` in favor of a `cloudmapNamespace: INamespace` parameter; that way, we can force resolution by ARN under the hood of the construct and never trigger the namespace duplication behavior. 

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*