feat(stepfunctions): support string and file definitions (#25932)
This change to AWS CDK will allow for the easier design of workflows by using the workflow studio. The visual representation of workflows is more attractive and easier to understand for developers and business people. Additionally, this change will address related issues in the CDK, improving the overall functionality and user experience. Adding the definitionBody and definitionSubstitutions props and the union-like class DefinitionBody will make it easier to define workflows using this tool by allowing users to define workflows in different ways, such as from an external file or a string.

Closes https://github.com/aws/aws-cdk/issues/8146
Refs https://github.com/aws/aws-cdk/issues/18880

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*