fix(ecr): autoDeleteImages fails on multiple repositories (#25964)
When setting `autoDeleteImages: true` for multiple repositories in the same stack, permissions to do the actual deleting only get added to the first one. This is because the policy statement is added inside of the `getOrCreateProvider` method, and that method ensures that the provider is only created once.

Instead, this adds the policy statement on the provider itself, regardless of whether it was created or referenced.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*