chore(cx-api): feature flags are documented incorrectly (#25994)
```ts
//  !!! IMPORTANT !!!
//
//  When you introduce a new flag, set its 'introducedIn.v2' value to the literal string
// 'V2·NEXT', without the dot.
//
//  DO NOT USE A VARIABLE. DO NOT DEFINE A CONSTANT. The actual value will be string-replaced at
//  version bump time.
```

Well, this proved to be difficult. Some places we supplied `V2·NEXT`, _with_ the dot. This PR manually updates everything back to when they were released, because we've missed the boat on the magic V2NEXT updates.

cc @corymhall for visibility, since you and I were the culprits and I was copying you :)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*