chore(app-staging-synthesizer): reduce the size of DefaultStagingStack template by 5KB (#26024)
Since `DefaultStagingStack` is a stack that creates staging resources, it is bound to the CloudFormation [limit](https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/cloudformation-limits.html) of 51KB. This means that there is an upper limit on how many ECR repositories (and in turn, unique images) are allowed in the CDK App. In an effort to increase this limit, this PR does the following:

- Remove path metadata to save ~2KB
- Remove `AWS::CDK::Metadata` construct to save ~3KB
- Insert a dummy `UsingDefaultStagingStack` construct to the stacks that _use_ the `DefaultStagingStack` for analytics purposes.

In all, this should decrease the size of the DefaultStagingStack template by ~5KB, or ~10% of the overall size allowed.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*