fix(elbv2): correct wrong timeout validation (#26031)
Following the update of Network Load Balancer (NLB) in November 2022, the range for setting HealthCheckTimeoutSeconds for NLB is now between 2 and 120 seconds. However, the CDK has yet to be updated to reflect this change.

https://aws.amazon.com/about-aws/whats-new/2022/11/elastic-load-balancing-capabilities-application-availability/
https://docs.aws.amazon.com/elasticloadbalancing/latest/network/target-group-health-checks.html

This PR addresses this by modifying the validation, enabling us to set the HealthCheckTimeoutSeconds for the NLB.

I have modified and added the following validation in this PR:
- Ensure that the HealthCheckTimeoutSeconds is between 2 and 120.
- Ensure that HealthCheckTimeoutSeconds is not greater than HealthCheckIntervalSeconds.


Closes #26023.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*