fix(cdk-lib): Pass lookupRoleArn to NestedStackSynthesizer (#26116)
NestedStack's synthesizer doesn't receive the lookupRoleArn from the parent stack synthesizer, so the NestedStack tries with local credentials (of the deployment account) instead of assuming a cross-account role (on the target account) as regular non-nested Stack would do.

This PR aims to add lookupRoleArn reference to the StackSynthesizer class and IStackSynthesizer, so it can be use on the NestedStack to explicitly set an IAM role in case of parent stack having one already defined, so CDK uses the role instead of local credentials.

Closes https://github.com/aws/aws-cdk/issues/25171.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*