feat(core): acknowledge annotation warning messages
This PR adds the ability to acknowledge annotation warning messages.
I've updated some of the annotations in the `rds.Cluster` construct to
show how this will work. After this is merged we can work on adding the
ids to all of the warning messages throughout the construct library.

The main motivation behind this is to allow people to set the `--strict`
mode to fail synthesis on warnings. Currently it is all or nothing, you
have to get rid of _all_ warnings to use `--strict`. With this feature
users will be able to `acknowledge` warnings saying that they are aware,
but it does not apply to them.

Since we want all warnings to now have an id this will deprecate the
`addWarning` method and adds a new `addWarningV2` method.

Since the acknowledgements and warnings are written as metadata, it is
possible to enhance this in the future to report on warnings and
acknowledgements.