chore: updates container-definition user jsdoc (#26231)
Update JSDOC for `user` to clarify that this is a `user string` and not restricted to a user name.

Reason for this change: When viewing the documentation from an IDE it is unclear that something other than a user name can be used for example 100:101 ie UID and GID.

I have not fully read the following:

[CONTRIBUTING GUIDE]: https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md
[DESIGN GUIDELINES]: https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md

This commit/pull request only changes JSDoc.

Closes #26230

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*