fix(pkglint): library creation generates incorrect package names (#26330)
Fairly certain that `options.packageName` is only used in the README example generation so there should be no other side-effects.

The reason why this wasn't caught earlier is because we turned `rosetta:strict` off. I am working on a separate PR to fix those issues. These are not the only modules with faulty examples. The rest will be fixed in #26331

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*