fix(s3): auto-delete-objects fails when bucket doesn't exist (#26433)
The AutoDeleteObjects Custom Resource should pass when the bucket doesn't exist.

With #16756 we introduced a safety check to only delete buckets that are marked for object-deletion.
This check would unintentionally bypass the special case to mark the CR deletion as successful when the bucket doesn't exist.
Additionally, with the upgrade to SDK v3 we need to change the check from `error.code` to check for the actual error instance.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*