fix(core): policy validation trace incorrect for larger constructs
Ran into an issue where the construct trace was incorrect in larger
projects, specifically where there are constructs that contain multiple
constructs.

To get the construct trace tree we first construct a new tree that only
contains the path to the terminal construct (the one with the trace). We
drop all the other constructs in the tree that don't relate. For
example, if we had a tree like:

```
->App
-->MyStage
--->MyStack1
---->MyConstruct
----->Resource
--->MyStack2
---->MyConstruct
----->Resource
--->MyStack3
---->MyConstruct
----->Resource
```

And we want to get a new tree for
`/App/MyStage/MyStack2/MyConstruct/Resource` it should look like this:

```
->App
-->MyStage
--->MyStack2
---->MyConstruct
----->Resource
```

We weren't correctly generating the new tree correctly and would always
end up with the tree being the first item in the list.

I've updated one of the tests, and also tested this on a more complex
application.