feat(cli): add diff message on the number of stacks with differences (#26297)
Reported issue with `diff` is that it treats the fail return status for cases when there are actual diffs, making it hard to know what happened with pipeline automations.

The proposed solution adds a logging statement with a similar format that is used for deploy (but here the total time is reported) specifying how many stacks have differences, as presented below. As a result, it will be possible to check in pipelines for this logging statement to correctly detect situation when there are no actual changes, when there are changes, and when there are failures, since on failures this statement will not be present:

Case of no changes:
✨  Number of stacks with differences: 0

Case of changes in 5 stacks:
✨  Number of stacks with differences: 5

Closes #10417.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*